Image classification package updates.
- Namespace from 'datalab_inception' to 'datalab_image_classification'.
- Move all flat methods (i.e. 'local_train') to Local and Cloud classes.
