+++ solutionbox/image_classification/datalab_image_classification/__init__.py
@@ -11,5 +11,5 @@
 # the License.
 
 
-from ._package import local_preprocess, cloud_preprocess, local_train, cloud_train, local_predict, \
-                      cloud_predict, local_batch_predict, cloud_batch_predict
+from ._local import Local
+from ._cloud import Cloud
