+++ solutionbox/image_classification/datalab_image_classification/_preprocess.py
